api/requirements.txt
azure-functions
fastapi==0.115.0
pydantic-settings==2.6.0

# Database - Cosmos DB
azure-cosmos>=4.7.0

# AI / ML
groq>=1.0.0
anthropic>=0.40.0

# PDF processing
pymupdf==1.27.2.2

# Utils
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.27.2
Pillow==11.0.0
